Fix for missing key functions that were overwriting user remappings by writing back the default mapping - Adds key mapping with a non-functioning key string "- Not Set - Name of KeyFun" - e.g. "- Not Set - Refresh"
This seemed very straightforward but I was getting some pretty strange behavior on testing. I tried to fix the wierdness without success and eventually threw away all the extra fixes and went back to my simple change which now seems to work! This needs testing!

Oh, I did make one important change other than the fix and that is to copy in the default mapping into the custom if the custom is empty. More about this and some UI thoughts in Issue #95.